requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1
django-fsm==2.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-fsm==1.0.4
factory-boy==3.2.1
Faker==14.2.0
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
